+++ src/main/java/net/geforcemods/securitycraft/recipe/LimitedUseKeycardRecipe.java
@@ -3,75 +3,32 @@ package net.geforcemods.securitycraft.recipe;
 import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.components.KeycardData;
 import net.geforcemods.securitycraft.items.KeycardItem;
-import net.minecraft.core.HolderLookup;
-import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.crafting.CraftingBookCategory;
-import net.minecraft.world.item.crafting.CraftingInput;
-import net.minecraft.world.item.crafting.CustomRecipe;
 import net.minecraft.world.item.crafting.RecipeSerializer;
-import net.minecraft.world.level.Level;
 
-public class LimitedUseKeycardRecipe extends CustomRecipe {
+public class LimitedUseKeycardRecipe extends CombineRecipe {
 	public LimitedUseKeycardRecipe(CraftingBookCategory craftingBookCategory) {
 		super(craftingBookCategory);
 	}
 
 	@Override
-	public boolean matches(CraftingInput inv, Level level) {
-		boolean hasNormalKeycard = false;
-		boolean hasLimitedUseKeycard = false;
-
-		for (int i = 0; i < inv.size(); ++i) {
-			ItemStack stack = inv.getItem(i);
-			Item item = stack.getItem();
-
-			if (item instanceof KeycardItem) {
-				if (item != SCContent.LIMITED_USE_KEYCARD.get()) {
-					if (hasNormalKeycard || stack.getOrDefault(SCContent.KEYCARD_DATA, KeycardData.DEFAULT).limited())
-						return false;
-
-					hasNormalKeycard = true;
-				}
-				else { //item is SCContent.LIMITED_USE_KEYCARD.get()
-					if (hasLimitedUseKeycard)
-						return false;
-
-					hasLimitedUseKeycard = true;
-				}
-			}
-			else if (!stack.isEmpty())
-				return false;
-		}
-
-		return hasNormalKeycard && hasLimitedUseKeycard;
+	public boolean matchesFirstItem(ItemStack stack) {
+		return stack.getItem() instanceof KeycardItem && !matchesSecondItem(stack) && !stack.getOrDefault(SCContent.KEYCARD_DATA, KeycardData.DEFAULT).limited();
 	}
 
 	@Override
-	public ItemStack assemble(CraftingInput inv, HolderLookup.Provider lookupProvider) {
-		ItemStack keycard = ItemStack.EMPTY;
-
-		for (int i = 0; i < inv.size(); ++i) {
-			ItemStack stack = inv.getItem(i);
-			Item item = stack.getItem();
-
-			if (item instanceof KeycardItem && item != SCContent.LIMITED_USE_KEYCARD.get()) {
-				keycard = stack.copy();
-				break;
-			}
-		}
-
-		if (keycard.isEmpty())
-			return ItemStack.EMPTY;
-
-		keycard.update(SCContent.KEYCARD_DATA, KeycardData.DEFAULT, data -> data.setLimitedAndUsesLeft(true, 0));
-		keycard.setCount(2);
-		return keycard;
+	public boolean matchesSecondItem(ItemStack stack) {
+		return stack.is(SCContent.LIMITED_USE_KEYCARD);
 	}
 
 	@Override
-	public boolean canCraftInDimensions(int width, int height) {
-		return width * height >= 2;
+	public ItemStack combine(ItemStack keycardToCopy, ItemStack limitedUseKeycard) {
+		ItemStack outputKeycard = keycardToCopy.copy();
+
+		outputKeycard.update(SCContent.KEYCARD_DATA, KeycardData.DEFAULT, data -> data.setLimitedAndUsesLeft(true, 0));
+		outputKeycard.setCount(2);
+		return outputKeycard;
 	}
 
 	@Override